+++ NEWS.md
@@ -51,3 +51,7 @@ Dimethyl (KR), Trimethyl (KR).
 
 * The function `PlotProteinCoverage()` now reports the coverage individually in
 each plot rather than the total protein coverage.
+
+* MQmetrics now is adapted to MaxQuant v.2.0.1.0, since the column names 
+are different than in MaxQuant v.1.6.17.0. MQmetrics will detect the MaxQuant 
+version used and read the columns accordingly.